#19 fix convert script for edge cases + download button

+++ ofm_transfer_data/convert_xls_to_csv.sh
@@ -1,26 +1,32 @@
 #!/bin/bash
-for FILE in *.xls; do
-    FILENAME=$(echo $FILE | awk '{split($0,a,"."); print a[1]}')
-    if [ -e "$FILENAME.csv" ]
-    then
-        echo "File already exists: $FILENAME.csv"
-    else
-        # convert to csv
-        libreoffice --headless --convert-to csv $FILE
 
-        # get start index of header line
-        START=$(grep -n Nr\\.,Spieltag $FILENAME.csv | awk '{split($0,a,":"); print a[1]}')
+shopt -s nullglob
+set -- *.xls
+if [ "$#" -gt 0 ]; then
+    for FILE in *.xls; do
+        FILENAME=$(echo $FILE | awk '{split($0,a,"."); print a[1]}')
+        if [ -e "$FILENAME.csv" ]
+        then
+            echo "File already exists: $FILENAME.csv"
+        else
+            # convert to csv
+            libreoffice --headless --convert-to csv $FILE
 
-        # drop first lines until start of header line
-        tail -n +$START "$FILENAME.csv" > "$FILENAME.tmp"
+            # get start index of header line
+            START=$(grep -n Nr\\.,Spieltag $FILENAME.csv | awk '{split($0,a,":"); print a[1]}')
 
-        # remove csv in order to append to it later
-        rm -f "$FILENAME.csv"
+            # drop first lines until start of header line
+            tail -n +$START "$FILENAME.csv" > "$FILENAME.tmp"
 
-        # convert from windows encoding to utf-8
-        iconv -f cp1252 -t utf-8  "$FILENAME.tmp" > "$FILENAME.csv"
+            # remove csv in order to append to it later
+            rm -f "$FILENAME.csv"
 
-        # delete temporary file
-        rm -f "$FILENAME.tmp"
-    fi
-done
+            # convert from windows encoding to utf-8
+            iconv -f cp1252 -t utf-8  "$FILENAME.tmp" > "$FILENAME.csv"
+
+            # delete stuffz
+            rm -f "$FILENAME.tmp"
+            rm -f $FILE
+        fi
+    done
+fi
